#1 #3 develop. + execution infos and first implement. of public server
The public server has two resources
- DataRequests (Requests for data from the health department to affected
persons)
- DataSubmission (Data submissions from the affected persons to the
health department)

Each resource provides one interface to the outside world (the public
interface for the apps - *Controller) and one interface to the inside
world (the interface for the second server on the side of the health
departments - *HdController).

The data is stored persistently.