#27 merges endpoints for contacts and events
Because contacts and events are always requested and transmitted
together for an index case, these features of IRIS, and thus the API
endpoints and DTOs, are merged.

The DataSubmissionProcessor is improved to have to write less code in
the subclasses repeatedly.

The query of the data in the dummy app is changed accordingly so that
contacts and events can be entered together for a transmission.