chore: refactors code and package structure to reduce cross-references (#70)
* chore: cleanup dtos of data_request part and separate this for submiss.

* chore: deletes old SORMAS integration to reduce the code to be adjusted

* chore: refactors code and package structure to reduce cross-references

* Entities of event and case requests completely separated and DB schema
adapted for this. There is now one table for EventDataRequest and one
for CaseDataRequest. The denormalization is accepted for a simpler data
model. The data processing is also only short-term and stream-like. The
data is only brought into a relationship in the respective specialist
application, so normalization brings no added value.
* The Features-Properties is removed from the DataRequests, because it
is no longer necessary with the separated types. This also removes the
table data_request_feature.
* Sorted out old and generated code of DTOs.
* Changed the package structure from DataRequest + DataSubmission as
root packages to Case + Events to reduce the number of cross-references
and thus provide a cleaner structure.
* Changed all DTOs to simple classes with Lombok.

After all, an incompatible change of the DB is accepted once again. This
is still ok in the current alpha phase and much easier.

* chore: removes swagger-annotations dependency and all uses of these ann.