chore: removes test RPCClientConfig class and fixes code smell
Hiding classes is bound to some conditions and can lead to code that is
difficult to understand.

The EJC used by Eclipse disallows two identical classes and does not
compile this code.

As it turned out during the search for alternatives, the config overlay
is currently not necessary. The code and the tests work without this
trick.