chore: sets special tags for release images to allow an automatic update of only releases
Adds image tags with `latest` for major and minor image version and sets
the tags for the releases (without -`latest`) only if it is not a
prerelease (default channel of semantic-release). This allows an
automatic update of only releases.

I left the docker-compose files with the release tags. So here is no
automatic update until the first release, if they are not changed
manually.

Refs iris-connect/iris-backlog#195