chore: improves the async mailing and makes it optional to avoid errors
The property `iris.client.mailing.active` is now available to enable
mail sending. Since clients are already installed without configuring
mailing, it is now disabled by default and can be enabled. If it is
active, the mandatory properties
`spring.mail.properties.recipient.event.data-received.email` and
`spring.mail.properties.recipient.case.data-received.email` are checked
at startup.

The code has been simplified in a few places. Checks that are useless
when the properties are mandatory have been removed. Also the
determination of the recipient was unified in one place and it is now
available via TemplatedEmail.