chore: implements logging of failed login attempts
The login attempts are stored in the database and old entries are
deleted cyclically. All durations and thresholds are adjustable via
properties.

When the threshold value is exceeded, the repeated false login is
logged. The threshold value is incremented by a configurable multiplier.

A waiting time until the next possible login attempt is also prepared.
This value is also increased via a configurable multiplier.

Refs iris-connect/iris-backlog#185