feat: Activates waiting time after login attempts and uses username instead the IP as reference.
The reference for a login attempt is no longer the remote IP, but the
username. This is the recommendation of OSWAP.

The waiting time is now taken into account before checking a login and a
401 is returned if it has not yet expired. The message of the response
then contains the timestamp until when it has to be waited.

Improves code with listener classes and LoginAttemptsService.

I must add permitAll for `/error` to avoid forbidden request if a login
fails.

Refs  iris-connect/iris-backlog#225