chore: Renames the property for the backend endpoint to IRIS_BACKEND-SERVICE_ENDPOINT
The old environment variable IRIS_LOCATION-SERVICE_ENDPOINT is taken into account continue. IRIS_BACKEND-SERVICE_ENDPOINT is the new variable and is used in all templates and configuration files.

It is now checked that one of the two variables is set. For the validation a meaningful own message was added.

PR #323