chore(build): removes unnecessary trigger
Following the Gitflow process, a PR to main should only initiate from the release or a bugfix (1.**, later 2.**) branch. For both origin branches, the tests are already executed, along with a build of the whole application. A PR on main, on the other hand, does not start a full build of the application, which is not necessary here.