chore: code cleanup
We pass the injected service objects with setServices to jsonrpc4j and it takes all the interfaces of the classes of these objects. With setServiceInterfaces we additionally declare the same interfaces again in our case.

setAllowMultipleInheritance is a very strange configuration. I guess it is supposed to allow multiple services to be registered with the same interface. However, the configuration should not be relevant since the list of interfaces is a set at that point in the code. Apart from that, the first matching implementation is then used and we should avoid this random behavior if it occurs. However, currently all services have their own dedicated interface, so the setting is irrelevant.