feat: A readable code is now used for the data request to an index case. This consists of four random words and an abbreviation for the department.
* Generates and uses a readable code for the data request to an index case. This is easier to transmit by phone. The code is used to generate the CAT and the DAT in IRIS client and the application equally. The token handling in the EPS has not changed.
* Uses the central configuration of the BS for configurations that are global and relevant for the readable token.
* Saves the correct announcement token with the request now.

+ code cleanup:
We pass the injected service objects with setServices to jsonrpc4j and it takes all the interfaces of the classes of these objects. With setServiceInterfaces we additionally declare the same interfaces again in our case.

setAllowMultipleInheritance is a very strange configuration. I guess it is supposed to allow multiple services to be registered with the same interface. However, the configuration should not be relevant since the list of interfaces is a set at that point in the code. Apart from that, the first matching implementation is then used and we should avoid this random behavior if it occurs. However, currently all services have their own dedicated interface, so the setting is irrelevant.

Co-authored-by: Jens Kutzsche <github@gebea.de>
Refs iris-connect/iris-backlog#245
PR #460