chore(Codescan): sets Trivy-Action back to master
The fact that old alarms were still displayed in the GH was not due to this action, but to the fact that old analyses were stored in the GH due to a change in the workflow. These had to be deleted via the API.

The checkout step is preserved even if the code is not really used. Without this step an error occurs at the `upload-sarif`.