fix: Avoids irrelevant but logged ClassNotFoundException. It is correct that the class iris.backend_service.configurations.CentralConfigurationException is missing.
Jsonrpc4j always tries to translate the exception from an error message, which it can get from the backend service in our case, into a same exception in the client. Since we do not want such a tight coupling, these classes do not exist in the client. Jsonrpc4j logs the ClassNotFoundException and then continues with a standard exception. These logs can be irritating.

PR #527