fix: Cleans the `.env(.sample)` from values that are prescribed for the environments. Admins in the HDs do not have to do anything with it now, instead the correct values are set internally based on the `IRIS_ENV`.
Set fix values for network specific configurations in setting files.

Cleans up the `.env.sample` and the deployment files for a coherent use of the configuration.

Adds Spring profiles for the EPS networks (`live` and `staging`) with specific configurations and which is selected by `IRIS_ENV`. These configurations are now removed from the deployment and deprecated in the `.env.sample`.

Uses YAML "macros" to avoid many duplications of configuration values.

An ADR is added for coherent use of the configuration for EPS.