chore: improves exception handling to largely use a consistent (close to Spring standard) response format and fixes the partially empty response bodies
The `GlobalControllerExceptionHandler` has been extended to generate a "proper" response with a body according to the Spring standard if no body is given. For this the implementation of `ErrorAttributes` is used, which generates a body according to `{"status":"...", "message":"...", "errors":"...",...}`. Exceptions are the `ResponseStatusException` and exceptions annotated with `@ResponseStatus`. These are still handled by their special exception resolvers.

Adds an extension to `DefaultErrorAttribute` that creates a unified representation for `Errors` and `ConstraintViolationException` and returns this with the `errors` property in the Spring default response body. The validation errors are mapped as a map of property name and a list of error messages `[field name : [violation message]]`.

In tests, an attempt is made to test all relevant exception scenarios and to check them for a proper response (especially body). However, in cases where the exception is thrown in a filter, no body is generated in the tests, but it is generated at the real runtime. This is because in the test the error handling via the (fictitious) call of `/error` and thus of `BasicErrorController` and the implementation of `ErrorAttributes` is not done.