feat: Health departments with IRIS can send messages to each other, encrypted and secured via the IRIS-Connect secure network. There is a central output and input for messages. Currently, only text messages are possible. In the near future, data sets as attachments will follow.
Adds a new module (package) for HD search. This is a client to use the HD search service of the backend services. With this service, we can search the responsible health department for a ZIP code or a city name. This feature is used to find a recipient for a message over the ZIP code or city.

Adds another module for the management of IRIS messages. This includes a web interface for the frontend and a JSON-RPC interface with client and server to transfer messages from a IRIS client to another client over the EPS network. The messages are only text messages at the moment. An extension with attachments and datasets will be done later.

Add views for message overview and details (iris-message-details view) with edit functionality (set isRead).

Co-authored-by: Jens Kutzsche <github@gebea.de>
Refs iris-connect/iris-backlog#256
Refs iris-connect/iris-backlog#259
PR #541