chore(Tests): separates scheduling configuration and disables this for tests
Running jobs is unnecessary and undesirable in tests, as it often leads to errors due to missing conditions (such as other systems). Test that test the `@Scheduled` methods should in turn also not rely on the concurrent indefinite execution of the method by Spring and ensure this otherwise.

Therefore, execution is disabled in tests to avoid errors caused by unnecessary and unintended execution.

Removes also a duplicated `@EnableScheduling` in `JWTWhitelistCleanup`. This blocked the other configuration from taking effect.

PR #699