feat(Messages): Messages can now be used to exchange guests of events between health departments. This makes it possible to transmit the guests received through a data request to the responsible department. The data can be transferred directly from the event overview to a message or can also be added to a message as an attachment. This is the beginning, more data types will follow.
A "plugin system" for the message payloads has been created in BFF, through which the modules (such as events, cases ...) can provide the functions for the payload.

This abstraction is used by the message module to select payloads, integrate them into the JSON (export them), display them at the receiver and import them.

As the first concrete implementation of a plugin, the transmission of events and their guests is implemented in the events module.

A "plugin system" has also been created in the FE, which is used to display the appropriate views for the selection of the payload depending on the payload type (discriminator). Thus a specific selection is embedded into the general message view. For the events, the views of the event overview are used in the FE by the plugin and supplemented by selection elements.

Co-authored-by: Jens Kutzsche <github@gebea.de>
PR #640