chore: improves mapper testing by avoiding error-prone mocking
Introduces an annotation (with the necessary helper classes) for mapper tests that can be used to implement a Spring test that makes error-prone mocking unnecessary with the Spring context. Thereby the Spring context is built specifically for the mappers with general base mappers and the Spring Conversion Service including all converters to avoid long execution times with the whole context.