chore: introduces message resolving for RPC exceptions, like with Spring and `ResponseStatusException`
An attempt is now made to resolve the message of an exception (property `message`) via the `MessageSourceAccessor`. If this is successful, the translated text is used. If not, the original error is used. This way the text does not have to be resolved every time an exception is created and the behavior is similar to `ResponseStatusException`.

As an example, the `IrisMessageBuilderEps` has been adapted.

PR #776