feat: The client backend now also supports the use of a refresh token, which can be used to extend the short validity of the authentication. This makes it more convenient to use, especially in conjunction with a two-factor authentication.
Login now returns a Refresh JWT in an HTTP-only cookie in addition to the authentication token in the response.

A JWT can be refreshed via the `/refreshtoken` path. This requires an expired but whitelisted JWT and a valid whitelisted refresh JWT (both in the corresponding cookie). Both must contain the same username. The response then contains an HTTP-only cookie with the new valid JWT.

For the paths `/login` and `/refreshtoken` the token cookie is now ignored. This avoids unnecessary errors due to the delivery of invalid stale cookies, and the cookies also play no role for these endpoints.

The logout also clears the refresh tokens and returns an empty token and refresh token.