feat: Users can now use two-factor authentication with time-based one-time password (TOTP). If it is enabled, a TOTP is expected and verified by a corresponding app after the conventional login. To set up the app, the user is displayed a QR code by IRIS. It is also possible for the admin to activate this mandatorily via environment variable. If a 2FA is expected but has not yet been finally configured for a user with a successful verification, the QR code is displayed after the successful conventional login and the verification is performed.
A secret for the MFA has been added to `UserAccount` and a flag indicating whether the transfer of the secret was completed successfully with verification. Furthermore, the `UserService` was supplemented with methods for working with the secret.

The `security.auth.db.mfa.option` property was added to specify how the MFA should be used. Possible values are: `ALWAYS, OPTIONAL_DEFAULT_TRUE, OPTIONAL_DEFAULT_FALSE, DISABLED`.

The new `MfAuthenticationConfigController` allows querying the configuration for the MFA via the `/mfa/config` path. Thus the FE can be adapted accordingly.

An optional step (depending on the MFA configuration) has been added to the authentication process. If the conventional login was successful, with active MFA a token is returned which only allows access to `/mfa/otp`. This is used to check an OTP to complete the authentication. Only after that a general JWT is returned.

The responses to authentication requests now contain a body with the property `authenticationStatus`, which contains the status of the authentication and can take the following values: `AUTHENTICATED, PRE_AUTHENTICATED_MFA_REQUIRED, PRE_AUTHENTICATED_ENROLLMENT_REQUIRED`.

The new annotation `@WithMockIrisUser` has been added to create a user with `authorities = "USER"`. This is necessary because an authority is now required for the accesses.

Refs iris-connect/iris-backlog#251