chore: moves the deletion of http only cookies to the GlobalControllerExceptionHandler
The http only cookies should only be deleted if an AuthenticationException occurs during login or refresh token (thus in the controller). To enable a token refresh that is done after a 401 response, the http only cookies must not be deleted when the AuthenticationException occurs in a filter, otherwise the basis for the refresh will be disposed of.