feat: For JSON-RPC calls (calls from EPS), the client name submitted by EPS is now used as user (if available). Thus, the metadata of records created via JSON-RPC now also contain a user as creator and it is easier to see by whom the data was created.
The client name is used as the username along with a prefix to make duplication less likely.

If a user does not yet exist for the client, then one is automatically created so that the references can be stored in the database. The users are assigned the new role 'ANONYMOUS'.

Users with the role 'ANONYMOUS' are filtered out in the repository when loading the users for the overview. These users represent internal technical users that are not managed by admins.

PR #826