feat: In the `.env` (see `.env.sample`) now the configuration for the mail dispatch can be done. With this it is now possible to send notifications when new data has been transferred to the IRIS client (at the moment implemented for the data of an event).
The composite files, the scripts and the `.env.sample` were extended by the env. variables for the mail dispatch and thus the notifications.

The properties in `application*.properties` have been corrected to more appropriate default values.

The templates were corrected because they used wrong placeholders. In addition, the representation of date information was improved in the process.

The properties `fix.sender.name` and `fix.sender` have been changed from `-` to `.` because otherwise there are errors when using them if they are set as ENV variables. The values will then not be found.

Adapts test to the changes.

Refs #557